this is the final folder structure revision. PLEASE LOOK AT COMMIT TO MATCH SO THERE ARE NO FURTHER ERRORS
