--- src/main/java/com/obliq/obliq/REPOS/CardsRepository.java
@@ -1,10 +1,0 @@
-package com.obliq.obliq.REPOS;
-
-import com.obliq.obliq.ENTITYS.Card;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CardsRepository extends CrudRepository<Card, Long> {
-
-}
