--- src/main/java/com/obliq/obliq/REPOS/CommentRepository.java
@@ -1,14 +1,0 @@
-package com.obliq.obliq.REPOS;
-
-import com.obliq.obliq.ENTITYS.Comment;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface CommentRepository extends CrudRepository<Comment, Long> {
-    List<Comment> findByUserId(Long id);
-    List<Comment> findByPostId(Long id);
-
-}
