--- src/main/java/com/obliq/obliq/REPOS/PostRepository.java
@@ -1,17 +1,0 @@
-package com.obliq.obliq.REPOS;
-
-import com.obliq.obliq.ENTITYS.Post;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface PostRepository extends CrudRepository<Post, Long> {
-    List<Post> findByUserId(Long id);
-
-    @Query("SELECT p FROM Post p ORDER BY date_created DESC")
-    List<Post> findByDateCreated();
-
-}
