--- src/main/java/com/obliq/obliq/REPOS/TribesRepository.java
@@ -1,9 +1,0 @@
-package com.obliq.obliq.REPOS;
-
-import com.obliq.obliq.ENTITYS.Tribe;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TribesRepository extends CrudRepository<Tribe, Long> {
-}
