--- src/main/java/com/obliq/obliq/REPOS/UserRepository.java
@@ -1,13 +1,0 @@
-package com.obliq.obliq.REPOS;
-
-import com.obliq.obliq.ENTITYS.User;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-    User findByUsername(String username);
-
-
-}
