--- src/main/java/com/obliq/obliq/SECURITY/UserDetailsLoader.java
@@ -1,27 +1,0 @@
-package com.obliq.obliq.SECURITY;
-
-import com.obliq.obliq.ENTITYS.User;
-import com.obliq.obliq.ENTITYS.UserWithRoles;
-import com.obliq.obliq.REPOS.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserDetailsLoader implements UserDetailsService {
-    private final UserRepository users;
-
-    public UserDetailsLoader(UserRepository users) {
-        this.users = users;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = users.findByUsername(username);
-        if (user == null) {
-            throw new UsernameNotFoundException("No user found for " + username);
-        }
-        return new UserWithRoles(user);
-    }
-}
