--- src/main/java/com/obliq/obliq/TESTING/TestingRepo.java
@@ -1,13 +1,0 @@
-package com.obliq.obliq.TESTING;
-
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.List;
-
-public interface TestingRepo extends CrudRepository <Testing, Long>{
-
-
-
-
-}
