+++ src/main/java/com/obliq/obliq/controllers/AuthenticationController.java
@@ -1,0 +1,12 @@
+package com.obliq.obliq.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AuthenticationController {
+    @GetMapping("/login")
+    public String showLoginForm() {
+        return "users/login";
+    }
+}
