+++ src/main/java/com/obliq/obliq/models/Post.java
@@ -1,0 +1,113 @@
+package com.obliq.obliq.models;
+
+
+import org.hibernate.annotations.CreationTimestamp;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import javax.persistence.*;
+import java.util.Date;
+import java.util.List;
+
+@Entity
+@Table(name="posts")
+public class Post{
+
+@Id @GeneratedValue
+private long id;
+
+@Column (name = "title", nullable = false, length = 50)
+private String title;
+
+@Column (name = "body", nullable = false, length = 500)
+private String body;
+
+@Column (name = "card_id", nullable = false)
+private long cardID;
+
+@CreationTimestamp
+@Temporal(TemporalType.TIMESTAMP)
+@Column(nullable = false)
+@DateTimeFormat(pattern="dd.MM.yyyy HH:mm:ss")
+private Date dateCreated;
+
+@ManyToOne
+@JoinColumn (name="user_id", referencedColumnName = "id")
+private User user;
+
+
+@OneToMany (cascade = CascadeType.ALL, mappedBy = "post")
+private List<Comment> comments;
+
+public Post(){}
+public Post(String title, String body, Date dateCreated, User user, List<Comment> comments, long cardID) {
+        this.title = title;
+        this.body = body;
+        this.dateCreated = dateCreated;
+        this.user = user;
+        this.comments = comments;
+        this.cardID = cardID;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getBody() {
+        return body;
+    }
+
+    public void setBody(String body) {
+        this.body = body;
+    }
+
+    public Date getDateCreated() {
+        return dateCreated;
+    }
+
+    public void setDateCreated(Date dateCreated) {
+        this.dateCreated = dateCreated;
+    }
+
+    public User getUser() {
+        return user;
+    }
+
+    public void setUser(User user) {
+        this.user = user;
+    }
+
+    public List<Comment> getComments() {
+        return comments;
+    }
+
+    public void setComments(List<Comment> comments) {
+        this.comments = comments;
+    }
+
+    public long getCardID() {
+        return cardID;
+    }
+
+    public void setCardID(long cardID) {
+        this.cardID = cardID;
+    }
+}
+
+/* --1 (name of the table you want to create----------------------------------------------------------------------------------------- */
+/* --2 (name of the entity object you are creating----------------------------------------------------------------------------------------- */
+/* --3 (title of the column heading----------------------------------------------------------------------------------------- */
+/* --4 (If this entity is owned by an Object----------------------------------------------------------------------------------------- */
+/* --5 (Object owners nickname----------------------------------------------------------------------------------------- */
+
