+++ src/main/java/com/obliq/obliq/repositories/TribesRepository.java
@@ -1,0 +1,9 @@
+package com.obliq.obliq.repositories;
+
+import com.obliq.obliq.models.Tribe;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TribesRepository extends CrudRepository<Tribe, Long> {
+}
