+++ src/main/java/com/obliq/obliq/repositories/UserRepository.java
@@ -1,0 +1,12 @@
+package com.obliq.obliq.repositories;
+
+import com.obliq.obliq.models.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Long> {
+    User findByUsername(String username);
+
+
+}
