+++ src/main/java/com/obliq/obliq/security/UserDetailsLoader.java
@@ -1,0 +1,27 @@
+package com.obliq.obliq.security;
+
+import com.obliq.obliq.models.User;
+import com.obliq.obliq.models.UserWithRoles;
+import com.obliq.obliq.repositories.UserRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserDetailsLoader implements UserDetailsService {
+    private final UserRepository users;
+
+    public UserDetailsLoader(UserRepository users) {
+        this.users = users;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = users.findByUsername(username);
+        if (user == null) {
+            throw new UsernameNotFoundException("No user found for " + username);
+        }
+        return new UserWithRoles(user);
+    }
+}
