+++ src/main/java/com/obliq/obliq/testing/TestingRepo.java
@@ -1,0 +1,11 @@
+package com.obliq.obliq.testing;
+
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface TestingRepo extends CrudRepository <Testing, Long>{
+
+
+
+
+}
