+++ src/main/resources/templates/users/signup.html
@@ -36,7 +36,7 @@
                         <label for="bio">bio</label>
                         <textarea type="text" th:field="*{bio}" class="form-control" id="bio"
                                   placeholder="Write a short bio to tell your tribe who you are"></textarea>
-                    </div>
+                    </div>kj
 
                     <div class="row">
                         <div class="col-md-6">
